[libcalamaresui] Paste TCP port is quint16
 - no sense in feeding in an int (and getting a compiler warning
   for narrowing) into something that by spec is a 16-bit value.

+++ src/libcalamaresui/utils/Paste.cpp
@@ -29,7 +29,7 @@ namespace CalamaresUtils
 {
 
 QString
-sendLogToPastebin( QObject* parent, const QString& ficheHost, int fichePort )
+sendLogToPastebin( QObject* parent, const QString& ficheHost, quint16 fichePort )
 {
     QString pasteUrlFmt = parent->tr( "Install log posted to:\n%1" );
     QFile pasteSourceFile( Logger::logFile() );
