+++ src/libcalamaresui/utils/Paste.h
@@ -19,6 +19,8 @@
 #ifndef UTILS_PASTE_H
 #define UTILS_PASTE_H
 
+#include <qglobal.h>  // for quint16
+
 class QObject;
 class QString;
 
@@ -29,7 +31,7 @@ namespace CalamaresUtils
  *
  * Returns the (string) URL that the pastebin gives us.
  */
-QString sendLogToPastebin( QObject* parent, const QString& ficheHost, int fichePort );
+QString sendLogToPastebin( QObject* parent, const QString& ficheHost, quint16 fichePort );
 
 }  // namespace CalamaresUtils
